Separate the CreateReviewItem component in its own file
